requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
langchain
langchain-classic
langchain-openai
